fixing

+++ Api/Documentation/GenerateTypescriptClient.cs
@@ -29,7 +29,7 @@ public static class GenerateTypescriptClient
 
         var lines = code.Split(new[] { Environment.NewLine }, StringSplitOptions.None).ToList();
         var startIndex = lines.FindIndex(l => l.Contains("export interface BaseDto"));
-        if (startIndex >= 0)
+        if (startIndex >= 0 && startIndex+4<=lines.Count)
             lines.RemoveRange(startIndex, 4); // Remove 3 lines (interface declaration and two properties)
 
         lines.Insert(0, "import { BaseDto } from 'ws-request-hook';");
